Switch to double arithmetic
Using ints was an oversight.

Doubles are now used for internal arithmetic and are nicely returned with up to 15 decimal places (but not trailing 0's)
Also included in support for doubles to be passed in the arguments

+++ src/calc.h
@@ -21,5 +21,5 @@
 struct Token {
 	char type[3];
 	char charVal;
-	long int intVal;
+	double intVal;
 };
